+++ tests/e2e_tests/conftest.py
@@ -77,6 +77,12 @@ def pytest_addoption(parser):
         default=False,
         help="Skip all confirmations",
     )
+    parser.addoption(
+        "--plugin-args-string",
+        action="store",
+        default="",
+        help="Custom plugin-specific CLI args to include in the deploy call.",
+    )
     # parser.addoption(
     #     "--plugin",
     #     action="store",
@@ -88,13 +94,14 @@ def pytest_addoption(parser):
 # Bundle these options into a single object.
 class CLIOptions:
     def __init__(
-        self, pkg_manager, pypi, automate_all, skip_confirmations, plugin_name
+        self, pkg_manager, pypi, automate_all, skip_confirmations, plugin_name, plugin_args_string, 
     ):
         self.pkg_manager = pkg_manager
         self.pypi = pypi
         self.automate_all = automate_all
         self.skip_confirmations = skip_confirmations
         self.plugin_name = plugin_name
+        self.plugin_args_string = plugin_args_string
 
 
 @pytest.fixture(scope="session")
@@ -105,6 +112,7 @@ def cli_options(request):
         automate_all=request.config.getoption("--automate-all"),
         skip_confirmations=request.config.getoption("--skip-confirmations"),
         plugin_name=request.config.getoption("--plugin"),
+        plugin_args_string=request.config.getoption("--plugin-args-string")
     )